2303-capstone-group-C/src/matching_service/requirements.txt
fastapi==0.95.1
python-decouple==3.8
redis==4.5.5
requests==2.31.0
uvicorn==0.22.0
